### 2025-07-03 #8535 @rouson
Here are a few suggestions:

[FTorch] contributors:
1. [Joe Wallwork](https://reviewers.joss.theoj.org/reviewers/3861)
2. [Dominic orchard](https://reviewers.joss.theoj.org/reviewers/7466)
3. [Christopher Edsall](https://reviewers.joss.theoj.org/reviewers/5724)

[ATHENA] contributor:
[Ned Thaddeus Taylor](https://github.com/nedtaylor)

[FTorch]: https://joss.theoj.org/papers/10.21105/joss.07602
[ATHENA]: https://joss.theoj.org/papers/10.21105/joss.06492
