### 2025-07-09 a4ec818
feat(demo): start ADON implementation
