+++ demo/app/infer-aerosol.f90
@@ -43,89 +43,96 @@ contains
 
   subroutine read_stats_and_perform_inference(path)
     character(len=*), intent(in) :: path
-    integer,          parameter :: num_inputs = 80, num_outputs = 31
-    character(len=*), parameter :: network_file_name = "model.json"
-    character(len=*), parameter :: inputs_tensor_file_name = "aerosol_input.nc"
-    double precision cube_root
-    double precision, allocatable, dimension(:,:) :: aerosol_data, input_components, output_components
-    type(tensor_statistics_t) input_stats, output_stats
-    type(unmapped_network_t(double_precision)) neural_network
+    integer,          parameter :: branch_net_inputs = 39, branch_net_outputs = 20
+    integer,          parameter :: trunk_net_inputs = 4, trunk_net_outputs = 20
+    character(len=*), parameter :: branch_file_name = "model_branch.json", trunk_file_name = "model_trunk.json"
+    character(len=*), parameter :: saved_data_file_name = "saved_data.nc"
+    double precision, allocatable, dimension(:) :: longitude, latitude, level, ymean, mean_X, std_X, mean_y, std_y
+    double precision, allocatable, dimension(:,:) :: cldfr_idx, basis, X_test, input_components, output_components
+    type(unmapped_network_t(double_precision)) branch_network, trunk_network
     integer i, j
 
-    input_stats  = read_tensor_statistics(path // "meanxp.txt", path // "stdxp.txt", num_inputs)  !for pre-processing normalization
-    output_stats = read_tensor_statistics(path // "meanyp.txt", path // "stdyp.txt", num_outputs) !for post-processing normalization
+    associate(saved_data_file => NetCDF_file_t(path // saved_data_file_name))
+      print '(a)', "Reading saved from " // path // saved_data_file_name
+      call saved_data_file%input("longitude", longitude)
+      call saved_data_file%input("latitude", latitude)
+      call saved_data_file%input("level", level)
+      call saved_data_file%input("cldfr_idx", cldfr_idx)
+      call saved_data_file%input("basis", basis)
+      call saved_data_file%input("ymean", ymean)
+      call saved_data_file%input("X_test", X_test)
+      call saved_data_file%input("mean_X", mean_X)
+      call saved_data_file%input("std_X", std_X)
+      call saved_data_file%input("mean_y", mean_y)
+      call saved_data_file%input("std_y", std_y)
+    end associate
 
-    associate(inputs_tensor_file => NetCDF_file_t(path // inputs_tensor_file_name))
-      print *,"Reading network inputs from ", inputs_tensor_file_name
-      call inputs_tensor_file%input("input_vars", aerosol_data)
-    end associate    
+    print *, "Reading the branch network from " // branch_file_name
+    branch_network = unmapped_network_t(double_precision_file_t(path // branch_file_name))
+    print *, "Reading the trunk network from " // trunk_file_name
+    trunk_network = unmapped_network_t(double_precision_file_t(path // trunk_file_name))
 
-    ! print*,'shape aerosol array = ',shape(aerosol_data) ! convert to an asssertion
-    
-    allocate(input_components(size(aerosol_data,2),size(aerosol_data,1)))
-    allocate(output_components(size(aerosol_data,2),num_outputs))
+    allocate(input_components(size(X_test,2),size(X_test,1)))
 
-    !$omp parallel do shared(aerosol_data,input_components,input_stats) private(i,j,cube_root)
     pre_process: &
-    do j = 1,size(aerosol_data,2)
-      do i = 1,size(aerosol_data,1)
[…]
