### 2025-07-09 6560898
fix(example): correct usage output
