+++ example/learn-exponentiation.F90
@@ -33,7 +33,7 @@ program learn_exponentiation
 
   if (len(final_network_file%string())==0) then
     error stop new_line('a') // new_line('a') // &
-      'Usage: fpm run --example train-polynomials --profile release --flag "-fopenmp" -- --output-file "<file-name>"'
+      'Usage: fpm run --example learn-exponentiation--profile release --flag "-fopenmp" -- --output-file "<file-name>"'
   end if
 
   block
