### 2025-07-14 0382a09
feat(demo): continue ADON implementation
