### 2025-07-14 da6d2f3
feat(demo): continue ADON implementation
