### 2025-07-15 0570e6b
chore(demo/setup): rm unneeded flags
