### 2025-07-15 3cda902
chore(demo/setup): rm unneeded flags
