### 2025-07-17 a304cf9
feat: compute trunk_inputs
