### 2025-07-31 c763e74
fix(neural_network_t): mk generic binding public
