### 2025-08-01 #8535 @espottesmith
COI check-in: I'm not currently affiliated with LBNL, but I was until recently (graduate student researcher until December 2023, affiliate researcher until December 2024). I have no relation to the authors/contributors (as far as I'm aware, anyways).

@crvernon, is it fair for me to take this on as editor?
