### 2025-08-01 7b0ed2b
doc(README): nagfor supported; describe llvm vers
