### 2025-08-01 b4d651d
doc(README): update ifx status - only 1 test fails
