### 2025-08-05 #8535 @espottesmith
Understood, @nedtaylor. I appreciate your quick response, in any case, and I hope you get a bit less busy soon!

Thanks for being willing to review, @jwallwork23! Getting started in a week is totally fine.
