### 2025-08-05 08ccbe0
chore(tests):update to Julienne 2.4.3,Assert 2.1.2
This commit
1. Switches all tests to test_diagnosis_t functions instead of
   the logical functions employed by Julienne 1.x,
2. Allows and tracks skipped tests in the "report" function as
   required in Julienne 2.x,
3. Leverages Julienne operator-based idioms in test functions, and
4. Removes all instances of Julienne's deprecated
   vector_function_strategy_t and vector_test_description_t types.
