+++ test/neural_network_test_m.F90
@@ -5,10 +5,18 @@ module neural_network_test_m
 
   ! External dependencies
   use kind_parameters_m, only : double_precision
-  use julienne_m, only : test_t, test_result_t, test_description_t, test_description_substring, string_t, file_t
-#ifdef __GFORTRAN__
-  use julienne_m, only : test_function_i
-#endif
+  use julienne_m, only : &
+      file_t &
+     ,operator(.all.) &
+     ,operator(.also.) &
+     ,operator(.approximates.) &
+     ,operator(.within.) &
+     ,string_t &
+     ,test_diagnosis_t &
+     ,test_t &
+     ,test_result_t &
+     ,test_description_t &
+     ,test_description_substring
 
   ! Internal dependencies
   use fiats_m, only : neural_network_t, tensor_t, metadata_t
@@ -35,35 +43,14 @@ contains
     type(test_result_t), allocatable :: test_results(:)
     type(test_description_t), allocatable :: test_descriptions(:)
 
-#ifndef __GFORTRAN__
     test_descriptions = [ &
        test_description_t("performing elemental inference with 1 hidden layer", elemental_infer_with_1_hidden_layer_xor_net) &
       ,test_description_t("performing elemental inference with 2 hidden layers", elemental_infer_with_2_hidden_layer_xor_net) &
-      ,test_description_t("converting a network with 2 hidden layers to and from JSON format", multi_hidden_layer_net_to_from_json)&
+      ,test_description_t("converting a network with 2 hidden layers to and from JSON format", multi_hidden_layer_net_to_from_json) &
       ,test_description_t("converting a network with varying-width hidden layers to/from JSON", varying_width_net_to_from_json) &
       ,test_description_t("performing inference with a network with hidden layers of varying width", infer_with_varying_width_net) &
       ,test_description_t("double-precision inference", double_precision_inference) &
     ]
-#else
-    procedure(test_function_i), pointer :: elemental_infer_1_ptr, elemental_infer_2_ptr, multi_hidden_ptr, vary_width_ptr, &
-      vary_width_infer_ptr, double_precision_inference_ptr
-
-    elemental_infer_1_ptr => elemental_infer_with_1_hidden_layer_xor_net
-    elemental_infer_2_ptr => elemental_infer_with_2_hidden_layer_xor_net
-    multi_hidden_ptr => multi_hidden_layer_net_to_from_json
-    vary_width_ptr => varying_width_net_to_from_json
-    vary_width_infer_ptr => infer_with_varying_width_net
-    double_precision_inference_ptr => double_precision_inference
-
-    test_descriptions = [ &
-       test_description_t("performing elemental inference with 1 hidden layer", elemental_infer_1_ptr) &
-      ,test_description_t("performing elemental inference with 2 hidden layers", elemental_infer_2_ptr) &
-      ,test_description_t("converting a network with 2 hidden layers to and from JSON format", multi_hidden_ptr) &
-      ,test_description_t("converting a network with varying-width hidden layers to/from JSON", vary_width_ptr) &
-      ,test_description_t("performing inference with varying-width hidden layers", vary_width_infer_ptr)  &
-      ,test_description_t("double-precision inference", double_precision_inference_ptr) &
-    ]
-#endif
     associate( &
       substring_in_subject => index(subject(), test_description_substring) /= 0, &
       substring_in_description => test_descriptions%contains_text(string_t(test_description_substring)) &
@@ -181,28 +168,29 @@ contains
     )   
   end function
 
-  function multi_hidden_layer_net_to_from_json() result(test_passes)
-    logical test_passes
+  function multi_hidden_layer_net_to_from_json() result(test_diagnosis)
+    type(test_diagnosis_t) test_diagnosis
[…]
