+++ test/tensor_test_m.f90
@@ -5,10 +5,18 @@ module tensor_test_m
 
   ! External dependencies
   use kind_parameters_m, only : double_precision 
-  use julienne_m, only : test_t, test_result_t, test_description_t, test_description_substring, string_t, file_t
-#ifdef __GFORTRAN__
-  use julienne_m, only : test_function_i
-#endif
+  use julienne_m, only : &
+     operator(.all.) &
+    ,operator(.also.) &
+    ,operator(.approximates.) &
+    ,operator(.within.) &
+    ,test_description_substring &
+    ,test_description_t &
+    ,test_diagnosis_t &
+    ,test_result_t &
+    ,test_t &
+    ,string_t &
+    ,file_t
 
   ! Internal dependencies
   use fiats_m, only : tensor_t
@@ -35,19 +43,9 @@ contains
     type(test_result_t), allocatable :: test_results(:)
     type(test_description_t), allocatable :: test_descriptions(:)
 
-#ifndef __GFORTRAN__
     test_descriptions = [ &
       test_description_t("double-precision construction and value extraction", double_precision_construction) &
     ]
-#else
-    procedure(test_function_i), pointer :: double_precision_construction_ptr
-      
-    double_precision_construction_ptr => double_precision_construction
-
-    test_descriptions = [ &
-      test_description_t("double-precision construction and value extraction", double_precision_construction_ptr) &
-    ]
-#endif
     associate( &
       substring_in_subject => index(subject(), test_description_substring) /= 0, &
       substring_in_description => test_descriptions%contains_text(string_t(test_description_substring)) &
@@ -57,15 +55,22 @@ contains
     test_results = test_descriptions%run()
   end function
 
-  function double_precision_construction() result(test_passes)
-    logical test_passes
+  function double_precision_construction() result(test_diagnosis)
+    type(test_diagnosis_t) test_diagnosis
     type(tensor_t(double_precision)) tensor
     double precision, parameter :: tolerance = 1.0D-12
     double precision, parameter :: values(*) = [1.000000000001D-12]
 
     tensor = tensor_t(values) ! this will fail to compile if no double_precision constructor exists and the 
     associate(tensor_values => tensor%values())
-      test_passes = kind(tensor_values) == double_precision .and. all(abs(tensor_values - values) < tolerance)
+      associate(kinds => kind(tensor_values))
+        test_diagnosis = &
+          test_diagnosis_t( &
+            test_passed        =  kinds == double_precision  &
+           ,diagnostics_string = "kind "// string_t(kinds) // " is not double precision" &
+          ) &
+          .also. (.all. (tensor_values .approximates. values .within. tolerance))
+      end associate
     end associate
   end function
 
