+++ fpm.toml
@@ -5,5 +5,5 @@ author = "Damian Rouson, Tan Nguyen, Jordan Welsman, David Torres, Brad Richards
 maintainer = "rouson@lbl.gov"
 
 [dependencies]
-assert = {git = "https://github.com/sourceryinstitute/assert", tag = "2.1.0"}
-julienne = {git = "https://github.com/berkeleylab/julienne", tag = "1.5.3"}
+assert = {git = "https://github.com/sourceryinstitute/assert", tag = "2.1.2"}
+julienne = {git = "https://github.com/berkeleylab/julienne", tag = "2.4.3"}

[…]
