### 2025-08-06 27a90b4
chore(fpm): simplify manifest
This commit removes unnecessary fields from the fpm.toml file in
order to simplify maintenance.
