+++ fpm.toml
@@ -1,8 +1,4 @@
 name = "fiats"
-version = "0.15.0"
-license = "see LICENSE.txt"
-author = "Damian Rouson, Tan Nguyen, Jordan Welsman, David Torres, Brad Richardson, Katherine Rasmussen, Federica Villani, Dan Bonachea"
-maintainer = "rouson@lbl.gov"
 
 [dependencies]
 assert = {git = "https://github.com/sourceryinstitute/assert", tag = "2.1.2"}
