### 2025-08-07 932b586
build(fpm): update to Assert 3.0.2
