+++ fpm.toml
@@ -1,5 +1,5 @@
 name = "fiats"
 
 [dependencies]
-assert = {git = "https://github.com/sourceryinstitute/assert", tag = "2.1.2"}
+assert = {git = "https://github.com/sourceryinstitute/assert", tag = "3.0.2"}
 julienne = {git = "https://github.com/berkeleylab/julienne", tag = "2.4.3"}
