### 2025-08-08 5a7bbca
chore: use julienne_assert in src/ and example
