### 2025-08-08 edc6990
refac: convert assertions to call_julienne_assert_
