### 2025-08-08 0eb1841
refac: convert assertions to call_julienne_assert_
