### 2025-08-08 d38ffb6
fix(example): add required associate nesting
