+++ example/learn-exponentiation.F90
@@ -19,9 +19,11 @@ contains
   elemental function y(x_tensor) result(a_tensor)
     type(tensor_t), intent(in) :: x_tensor
     type(tensor_t) a_tensor
-    associate(x => x_tensor%values(), suffient_input => (ubound(x,1) .isAtLeast. 7) .also. (lbound(x,1) .isAtMost. 2))
-      call_julienne_assert(suffient_input)
-      a_tensor = tensor_t([x(1)**2, x(2)**3, x(3)**4, x(4)**4, x(5)**3, x(6)**2])
+    associate(x => x_tensor%values())
+      associate(suffient_input => (ubound(x,1) .isAtLeast. 7) .also. (lbound(x,1) .isAtMost. 2))
+        call_julienne_assert(suffient_input)
+        a_tensor = tensor_t([x(1)**2, x(2)**3, x(3)**4, x(4)**4, x(5)**3, x(6)**2])
+      end associate
     end associate
   end function
 
