+++ test/tensor_map_test_m.F90
@@ -16,9 +16,6 @@ module tensor_map_test_m
     ,test_result_t &
     ,test_t
   use fiats_m, only : tensor_map_t, tensor_t
-#ifdef __GFORTRAN__
-  use julienne_m, only : test_function_i
-#endif
 
   ! Internal dependencies
   use tensor_map_m, only : tensor_map_t
@@ -45,21 +42,10 @@ contains
     type(test_result_t), allocatable :: test_results(:)
     type(test_description_t), allocatable :: test_descriptions(:)
 
-#ifndef __GFORTRAN__
     test_descriptions = [ & 
       test_description_t("component-wise construction followed by conversion to and from JSON", write_then_read_tensor_map), &
       test_description_t("mapping to and from the unit interval as an identity transformation", map_to_from_training_range) &
     ]
-#else
-    procedure(test_function_i), pointer :: check_write_then_read_ptr, check_map_to_from_range_ptr 
-    check_write_then_read_ptr   => write_then_read_tensor_map
-    check_map_to_from_range_ptr => map_to_from_training_range
-
-    test_descriptions = [ &
-      test_description_t("component-wise construction followed by conversion to and from JSON", check_write_then_read_ptr), &
-      test_description_t("mapping to and from the unit interval as an identity transformation", check_map_to_from_range_ptr) &
-    ]
-#endif
     associate( &
       substring_in_subject => index(subject(), test_description_substring) /= 0, &
       substring_in_description => test_descriptions%contains_text(string_t(test_description_substring)) &
