+++ test/training_configuration_test_m.F90
@@ -13,9 +13,6 @@ module training_configuration_test_m
     ,test_diagnosis_t &
     ,test_result_t &
     ,test_t
-#ifdef __GFORTRAN__
-  use julienne_m, only : test_function_i
-#endif
 
   ! Internal dependencies
   use training_configuration_m, only : training_configuration_t
@@ -41,22 +38,11 @@ contains
     type(test_result_t), allocatable :: test_results(:)
     type(test_description_t), allocatable :: test_descriptions(:)
 
-#ifndef __GFORTRAN__
     test_descriptions = [ & 
       test_description_t( &
         string_t("component-wise construction followed by conversion to and from JSON"), &
         construct_and_convert_to_and_from_json) & 
     ]
-#else
-    procedure(test_function_i), pointer :: check_write_then_read_ptr
-    check_write_then_read_ptr => construct_and_convert_to_and_from_json
-
-    test_descriptions = [ &
-      test_description_t( &
-        string_t("component-wise construction followed by conversion to and from JSON"), &
-        check_write_then_read_ptr) &
-    ]
-#endif
     associate( &
       substring_in_subject => index(subject(), test_description_substring) /= 0, &
       substring_in_description => test_descriptions%contains_text(string_t(test_description_substring)) &

[…]
