### 2025-08-08 5533d0d
fix(example): add required associate nesting
