### 2025-08-08 b52d314
feat: rm gfortran workaround
This commit removes the test-suite workarounds for a Fortran 2008
feature that was added in gfortran 14.3 and later.
