### 2025-08-08 7cc48ac
fix(tensor_map_test): typo
