### 2025-08-12 #8785 @editorialbot
[REVIEW]: Fiats: Functional inference and training for surrogates
**Submitting author:** @rouson (<a href="http://orcid.org/0000-0002-2344-868X">Damian Rouson</a>)
**Repository:** https://github.com/berkeleylab/fiats
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 1.2.0
