### 2025-08-21 060bdb1
updte infer-aerosol
