+++ demo/cols.txt
@@ -1,0 +1,7 @@
+tensor([[[-1.2802e+00,  5.8679e-01,  4.1886e-01,  ...,  4.4384e-11,
+           2.1691e-08, -2.3857e-03],
+         [-5.0868e-01, -2.0909e+00,  4.3652e-03,  ...,  1.0448e-11,
+           5.5577e-09,  6.4837e-03],
+         [ 1.3825e+00,  6.3053e-02, -8.7317e-01,  ...,  6.7257e-04,
+           2.5975e+00,  2.8966e-03],
+Last col:  [-2.3857e-03,  6.4837e-03, 2.8966e-03,…]
