### 2025-08-22 e3ed4ec
Fix bullet rendering in README
