### 2025-08-22 #8785 @jwallwork23
I had difficulty installing `flang-new` (the build crashed twice, killing the terminal both times) and didn't have new enough versions of `ifx` or `nagfor` available but eventually managed to get `ifx` 2025.2.1 installed. As mentioned in the README, all tests pass except one.

Here's some initial feedback on the paper:

General feedback:
	* Paper is far over the 1000 word limit and needs cutting back significantly.
* Statement of need section:
	* Whilst there is a statement of need, I don't feel this is stated clearly enough.
	* Long discussion related to Fortran standard can be cut down.
* State of the field section:
	* Is it worth mentioning https://github.com/CrayLabs/SmartSim? Whilst it's a more general tool, we've found that some teams use it to couple PyTorch into Fortran codes.
	* Some teams also use https://github.com/ylikx/forpy to give a Fortran code access to a Python runtime with PyTorch in it.
	* Please cite the [FTorch JOSS paper](https://joss.theoj.org/papers/10.21105/joss.07602) rather than the repo URL.
	* Drop "The other mentioned projects have most-recent commits no older than two months as of June 2025." as unnecessary.
* Recent research section:
	* The class diagrams take up a lot of space but I feel don't add much to the paper. Many of them are very simple. If you feel the class diagrams are needed, could they be represented as code within the text blocks rather than floating figures?

I'll do the software side of the review separately and open issues as I come to them.
