+++ README.md
@@ -199,17 +199,17 @@ Documentation
 Please see our [GitHub Pages site] for Hypertext Markup Languge (HTML) documentation generated by [`ford`] or generate documentation locally by installing `ford` and executing `ford ford.md`.
 
 ### UML
-Please see the `doc/uml` subdirectory for Unified Modeling Language (UML) diagrams such as a comprehensive Fiats [class diagram] with human-readable [Mermaid] source that renders graphically when opened by browsing to the document on GitHub.
+Please see the [`doc/uml`](https://github.com/BerkeleyLab/fiats/blob/main/doc/uml) subdirectory for Unified Modeling Language (UML) diagrams such as a comprehensive Fiats [class diagram] with human-readable [Mermaid] source that renders graphically when opened by browsing to the document on GitHub.
 
 [Building and testing]: #building-and-testing
 [Caffeine]: https://go.lbl.gov/caffeine
-[class diagram]: ./doc/uml/class-diagram.md
+[class diagram]: https://github.com/BerkeleyLab/fiats/blob/main/doc/uml/class-diagram.md
 [Documentation]: #documentation
-[demo]: demo
+[demo]: https://github.com/BerkeleyLab/fiats/blob/main/demo
 [E3SM]: https://e3sm.org
-[example]: example
-[example/print-training-configuration.F90]: example/print-training-configuration.F90
-[example/concurrent-inferences]: example/concurrent-inferences.f90
+[example]: https://github.com/BerkeleyLab/fiats/blob/main/example
+[example/print-training-configuration.F90]: https://github.com/BerkeleyLab/fiats/blob/main/example/print-training-configuration.F90
+[example/concurrent-inferences]: https://github.com/BerkeleyLab/fiats/blob/main/example/concurrent-inferences.f90
 [`ford`]: https://github.com/Fortran-FOSS-Programmers/ford
 [`fpm`]: https://github.com/fortran-lang/fpm
 [Getting Started]: #getting-started
@@ -225,4 +225,4 @@ Please see the `doc/uml` subdirectory for Unified Modeling Language (UML) diagra
 [Overview]: #overview
 [ROCm fork]: https://github.com/ROCm/llvm-project
 [rojff]: https://gitlab.com/everythingfunctional/rojff
-[saturated_mixing_ratio_m]: example/supporting-modules/saturated_mixing_ratio_m.f90
+[saturated_mixing_ratio_m]: https://github.com/BerkeleyLab/fiats/blob/main/example/supporting-modules/saturated_mixing_ratio_m.f90
