### 2025-08-22 BerkeleyLab/inference-engine#273 @joewallwork
[JOSS review]: Dependencies list
Towards https://github.com/openjournals/joss-reviews/issues/8785.

The JOSS review checklist includes
> Is there a clearly-stated list of dependencies?

Please could you provide such a list in the README, or somewhere else clearly visible?
