### 2025-08-29 BerkeleyLab/fiats#272 @rouson
@jwallwork23 thanks for submitting this issue.  Several of the co-authors have developed contributing guidelines for other repositories (e.g, [Caffeine](https://go.lbl.gov/caffeine)) so we will adopt/adapt those guidelines and add them here.
