### 2025-08-29 61c792a
Add CONTRIBUTING.md first draft
