### 2025-08-29 b2da2d1
Update CONTRIBUTING.MD
Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>
