### 2025-08-29 0340683
Update CONTRIBUTING.MD
Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>
