### 2025-08-29 502fa50
Update CONTRIBUTING.MD
Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>
