### 2025-08-29 0827efa
Update CONTRIBUTING.MD
Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>
