### 2025-08-29 17d2efc
Update LICENSE.txt
Update copyright date, fix spelling error
