+++ LICENSE.txt
@@ -1,6 +1,6 @@
-*** Licence Agreement ***
+*** License Agreement ***
 
-inference-engine Copyright (c) 2022, The Regents of the University of
+Fiats Copyright (c) 2022-2025, The Regents of the University of
 California, through Lawrence Berkeley National Laboratory (subject to
 receipt of any required approvals from the U.S. Dept. of Energy). 
 All rights reserved.
