### 2025-08-29 BerkeleyLab/fiats#273 @bonachea
Resolved in #275
