### 2025-08-29 439af53
Update contributors list in CONTRIBUTING.md
