### 2025-09-02 BerkeleyLab/fiats#276 @niccolozanotti
[JOSS Review] `flang` 21 compiler flag to parallelize `do concurrent`
In the README/documentation you suggest enabling the CPU parallelization of batch inferences via `flang-new` by passing the `-fdo-concurrent-parallel=host` flag along with `-fopenmp`.
However, this raises
```
flang-21: error: unknown argument: '-fdo-concurrent-parallel=host'
```
Perhaps this was the correct flag in an older versions of flang?

I was able to run it by passing `-fdo-concurrent-to-openmp=host` as described in the official [llvm docs](https://flang.llvm.org/docs/DoConcurrentConversionToOpenMP.html).

---
Towards https://github.com/openjournals/joss-reviews/issues/8785.
