### 2025-09-02 71efa9b
doc(README): fix auto-parallelization build info
