### 2025-09-02 BerkeleyLab/fiats#276 @rouson
@niccolozanotti thanks for submitting this.  Commit 71efa9b97a8256bd19c813884d73eeb35f0db24c fixes this issue.
