### 2025-09-04 #8785 @espottesmith
@jwallwork23 @niccolozanotti

Have y'all had a chance to review the most recent changes (I see the additional now-closed issue by @niccolozanotti above)?
