### 2025-09-05 #8785 @jwallwork23
> Have y'all had a chance to review the most recent changes (I see the additional now-closed issue by [@niccolozanotti](https://github.com/niccolozanotti) above)?

I followed those changes as they went through, yes. I can't tell if the paper has been updated based on my feedback yet so haven't taken another look at that yet.
