### 2025-09-05 #8785 @niccolozanotti
@espottesmith  I was able to to make some headway but I was slowed down by some issues I encountered while handling the different compilers to test the software as described. I have currently done some testing using llvm's `flang@21` on a macOS machine and with intel's `ifx` `2025.2.1` on a linux machine getting the expected results.
Oddly enough, I ran into some run errors when testing specific examples/demos on one machine and not the other. I'll investigate further and open related issues on the repository if needed.

One thing I particularly appreciate is the use of the newly introduced features of the language by the recent standards.

In terms of the manuscript, I share many of the points that @jwallwork23 has already mentioned, so I won't double down on all of those. I just want to further suggest to shorten the length by cutting down a bit the discussion of Fortran language standards and to consider removing some of the diagrams beyond the first one, or to otherwise state why you feel like those are worth keeping.

This piece of software constitutes a great addition to the open source tools for doing deep learning in Fortran and I am myself eager to use it.
