### 2025-09-30 154b9e8
chore: add compiler/version language suppor macros
