### 2025-09-30 ba947fb
feat(default_real_learn): default(none) locality
