+++ src/fiats/neural_network_s.F90
@@ -4,6 +4,7 @@
 #include "julienne-assert-macros.h"
 #include "assert_macros.h"
 #include "compound_assertions.h"
+#include "language-support.F90"
 
 submodule(neural_network_m) neural_network_s
   use assert_m
@@ -74,12 +75,16 @@ contains
       end block
 #endif
 
-      feed_forward: &
-      do l = input_layer+1, output_layer
-        associate(z => matmul(w(1:n(l),1:n(l-1),l), a(1:n(l-1),l-1)) + b(1:n(l),l))
-          a(1:n(l),l) = self%activation_%evaluate(z)
-        end associate
-      end do feed_forward
+      block
+        integer l
+
+        feed_forward: &
+        do l = input_layer+1, output_layer
+          associate(z => matmul(w(1:n(l),1:n(l-1),l), a(1:n(l-1),l-1)) + b(1:n(l),l))
+            a(1:n(l),l) = self%activation_%evaluate(z)
+          end associate
+        end do feed_forward
+      end block
 
 #ifndef _CRAYFTN
       associate(normalized_outputs => tensor_t(a(1:n(output_layer), output_layer)))
@@ -120,12 +125,16 @@ contains
       end block
 #endif
 
-      feed_forward: &
-      do l = input_layer+1, output_layer
-        associate(z => matmul(w(1:n(l),1:n(l-1),l), a(1:n(l-1),l-1)) + b(1:n(l),l))
-          a(1:n(l),l) = self%activation_%evaluate(z)
-        end associate
-      end do feed_forward
+      block
+        integer l
+
+        feed_forward: &
+        do l = input_layer+1, output_layer
+          associate(z => matmul(w(1:n(l),1:n(l-1),l), a(1:n(l-1),l-1)) + b(1:n(l),l))
+            a(1:n(l),l) = self%activation_%evaluate(z)
+          end associate
+        end do feed_forward
+      end block
 
 #ifdef _CRAYFTN
       block
@@ -806,7 +815,7 @@ contains
   end procedure
 
   module procedure default_real_learn
-    integer l, batch, mini_batch_size, pair
+    integer batch, mini_batch_size, pair
     type(tensor_t), allocatable :: inputs(:), expected_outputs(:)
 
     call_assert_consistency(self)
@@ -850,11 +859,11 @@ contains
               real, allocatable :: pair_cost(:)
               if (present(cost)) allocate(pair_cost(mini_batch_size))
 
-#if F2023_LOCALITY
+#if defined(F2023_LOCALITY)
               iterate_through_batch: &
-              do concurrent (pair = 1:mini_batch_size) default(none) local(a,z,delta) reduce(+: dcdb, dcdw)
-
-#elif F2018_LOCALITY
+              do concurrent (pair = 1:mini_batch_size) default(none)  local(a,z,delta) reduce(+: dcdb, dcdw) &
+                shared(self, inputs, output_layer, n, w, b, cost, expected_outputs, pair_cost)
+#elif defined(F2018_LOCALITY)
 
               reduce_gradients: &
               block
@@ -887,11 +896,14 @@ contains
 
                     a(1:self%num_inputs(), input_layer) = inputs(pair)%values()
 
-                    feed_forward: &
-                    do l = 1,output_layer
-                      z(1:n(l),l) = matmul(w(1:n(l),1:n(l-1),l), a(1:n(l-1),l-1)) + b(1:n(l),l) ! z_j^l =  sum_k(w_jk^{l} a_k^{l-1}) + b_j^l
[…]
