### 2025-10-18 e49d751
chore: update demo to Julienne 3.2.1 assertions
This commit replaces all occurrences of the `call_assert` and
`call_assert_diagnose` macros with either the `call_julienne_assert`
macro or the `call julienne_assert` subroutine invocation.  Doing so
allows for rich, automatically constructed diagnostic information
provided by Julienne idioms.  Where desirable, using calling the
directly facilitates insertion of custom diagnostic information.
