+++ demo/src/histogram_s.F90
@@ -1,7 +1,8 @@
 ! Copyright (c), The Regents of the University of California
 ! Terms of use are as specified in LICENSE.txt
 
-#include "assert_macros.h"
+#include "julienne-assert-macros.h"
+
 
 submodule(histogram_m) histogram_s
   use assert_m
@@ -76,8 +77,8 @@ contains
         block 
           integer h, b ! histogram number, bin number
 
-          call_assert(num_bins > 0)
-          call_assert_describe(all(histograms(1)%num_bins() == [(histograms(h)%num_bins() , h=1,size(histograms))]), "histogram_s(to_file): uniform number of bins")
+          call_julienne_assert(num_bins > 0)
+          call_julienne_assert(all(histograms(1)%num_bins() == [(histograms(h)%num_bins() , h=1,size(histograms))]))
             
           allocate(columns(num_bins))
           do b = 1, num_bins
@@ -97,7 +98,7 @@ contains
   pure function normalize(x, x_min, x_max) result(x_normalized)
     real, intent(in) :: x(:,:,:,:), x_min, x_max
     real, allocatable :: x_normalized(:,:,:,:)
-    call_assert_diagnose(x_min/=x_max, "histogram_m(normalize): x_min/=x_max", intrinsic_array_t([x_min, x_max]))
+    call_julienne_assert(x_min/=x_max)
     x_normalized = (x - x_min)/(x_max - x_min)
   end function
 
@@ -153,7 +154,7 @@ contains
       end associate
 #ifdef ASSERTIONS
       associate(binned => sum(bin_count))
-        call_assert_diagnose(cardinality == binned, "histogram_s(construct): lossless binning", intrinsic_array_t([cardinality, binned]))
+        call_julienne_assert(cardinality .equalsExpected. binned)
       end associate
 #endif
     end associate
