+++ test/tensor_names_test.f90
@@ -8,7 +8,6 @@ module tensor_names_test_m
   use julienne_m, only : &
     string_t &
    ,test_description_t &
-   ,test_description_substring &
    ,test_diagnosis_t &
    ,test_result_t &
    ,test_t 
@@ -34,24 +33,15 @@ contains
   end function
 
   function results() result(test_results)
+    type(tensor_names_test_t) tensor_names_test
     type(test_result_t), allocatable :: test_results(:)
-    type(test_description_t), allocatable :: test_descriptions(:)
 
-    test_descriptions = [ & 
+    test_results = tensor_names_test%run([ & 
       test_description_t("component-wise construction followed by conversion to and from JSON", write_then_read_tensor_names) &
-    ]
-    associate( &
-      substring_in_subject => index(subject(), test_description_substring) /= 0, &
-      substring_in_description => test_descriptions%contains_text(string_t(test_description_substring)) &
-    )
-      test_descriptions = pack(test_descriptions, substring_in_subject .or. substring_in_description)
-    end associate
-    test_results = test_descriptions%run()
+    ])
   end function
 
-  function write_then_read_tensor_names() result(test_diagnosis)
-    type(test_diagnosis_t) test_diagnosis
-
+  type(test_diagnosis_t) function write_then_read_tensor_names() result(test_diagnosis)
     associate( &
       from_components => tensor_names_t( &
         inputs  = [string_t("qc"), string_t("qv"), string_t("pressure")], &
