### 2025-10-19 4e7ee4f
chore(assertions): update to Julienne 3.2.1 forms
This commit finishes updating the assertions in the src, example,
demo, and test subdirectories using Julienne 3.2.1 idioms.
