+++ src/fiats/hyperparameters_s.F90
@@ -1,10 +1,10 @@
 ! Copyright (c) 2023-2025, The Regents of the University of California
 ! Terms of use are as specified in LICENSE.txt
 
-#include "assert_macros.h"
+#include "julienne-assert-macros.h"
 
 submodule(hyperparameters_m) hyperparameters_s
-  use assert_m
+  use julienne_m, only : call_julienne_assert_
   implicit none
 
   character(len=*), parameter :: mini_batches_key  = "mini-batches"
@@ -29,7 +29,7 @@ contains
 
     real, parameter :: tolerance = 1.E-08
 
-    call_assert(allocated(lhs%optimizer_) .and. allocated(rhs%optimizer_))
+    call_julienne_assert(allocated(lhs%optimizer_) .and. allocated(rhs%optimizer_))
 
     lhs_equals_rhs = &
       lhs%mini_batches_ == rhs%mini_batches_ .and. &
@@ -42,7 +42,7 @@ contains
 
     double precision, parameter :: tolerance = 1.D-15
 
-    call_assert(allocated(lhs%optimizer_) .and. allocated(rhs%optimizer_))
+    call_julienne_assert(allocated(lhs%optimizer_) .and. allocated(rhs%optimizer_))
 
     lhs_equals_rhs = &
       lhs%mini_batches_ == rhs%mini_batches_ .and. &
@@ -67,7 +67,7 @@ contains
       end if
     end do
 
-    call_assert(hyperparameters_key_found)
+    call_julienne_assert(hyperparameters_key_found)
   end procedure
 
   module procedure double_precision_from_json
@@ -86,7 +86,7 @@ contains
       end if
     end do
 
-    call_assert(hyperparameters_key_found)
+    call_julienne_assert(hyperparameters_key_found)
   end procedure
 
   module procedure default_real_to_json
