### 2025-10-19 a972878
fix(time_data_s):ignore leading space in assertion
