### 2025-10-20 a81f3ef
feat(assertions): convert logical to idiomatic
