+++ src/neural_network_s.F90
@@ -2,12 +2,10 @@
 ! Terms of use are as specified in LICENSE.txt
 
 #include "julienne-assert-macros.h"
-#include "assert_macros.h"
 #include "compound_assertions.h"
 
 submodule(neural_network_m) neural_network_s
-  use assert_m
-  use julienne_m, only : call_julienne_assert_ , operator(.all.), operator(.equalsExpected.)
+  use julienne_m, only : call_julienne_assert_ , operator(.all.), operator(.equalsExpected.), operator(.expect.)
   use double_precision_string_m, only : double_precision_string_t
   use kind_parameters_m, only : double_precision
   use layer_m, only : layer_t
@@ -62,17 +60,9 @@ contains
 
       allocate(a(maxval(n), input_layer:output_layer))
 
-#ifndef _CRAYFTN
       associate(normalized_inputs => self%input_map_%map_to_training_range(inputs))
         a(1:n(input_layer),input_layer) = normalized_inputs%values()
       end associate
-#else
-      block
-        type(tensor_t) normalized_inputs
-        normalized_inputs = self%input_map_%map_to_training_range(inputs)
-        a(1:n(input_layer),input_layer) = normalized_inputs%values()
-      end block
-#endif
 
       feed_forward: &
       do l = input_layer+1, output_layer
@@ -81,17 +71,9 @@ contains
         end associate
       end do feed_forward
 
-#ifndef _CRAYFTN
       associate(normalized_outputs => tensor_t(a(1:n(output_layer), output_layer)))
         outputs = self%output_map_%map_from_training_range(normalized_outputs)
       end associate
-#else
-      block
-        type(tensor_t) :: normalized_outputs
-        normalized_outputs = tensor_t(a(1:n(output_layer), output_layer))
-        outputs = self%output_map_%map_from_training_range(normalized_outputs)
-      end block
-#endif
 
     end associate
 
@@ -108,17 +90,9 @@ contains
 
       allocate(a(maxval(n), input_layer:output_layer))
 
-#ifndef _CRAYFTN
       associate(normalized_inputs => self%input_map_%map_to_training_range(inputs))
         a(1:n(input_layer),input_layer) = normalized_inputs%values()
       end associate
-#else
-      block
-        type(tensor_t) normalized_inputs
-        normalized_inputs = self%input_map_%map_to_training_range(inputs)
-        a(1:n(input_layer),input_layer) = normalized_inputs%values()
-      end block
-#endif
 
       feed_forward: &
       do l = input_layer+1, output_layer
@@ -127,19 +101,9 @@ contains
         end associate
       end do feed_forward
 
-#ifdef _CRAYFTN
-      block
-        type(tensor_t) :: normalized_outputs
-        normalized_outputs = tensor_t(a(1:n(output_layer), output_layer))
-#else
       associate(normalized_outputs => tensor_t(a(1:n(output_layer), output_layer)))
-#endif
         outputs = self%output_map_%map_from_training_range(normalized_outputs)
-#ifdef _CRAYFTN
-      end block
-#else
       end associate
-#endif
 
     end associate
 
@@ -147,9 +111,7 @@ contains
 
   module procedure default_real_consistency
 
-    call_assert(allocated(self%weights_))
-    call_assert(allocated(self%biases_))
-    call_assert(allocated(self%nodes_))
+    call_julienne_assert(.all. (.expect. [allocated(self%weights_), allocated(self%biases_), allocated(self%nodes_)]))
 
[…]
