+++ src/layer_s.F90
@@ -2,11 +2,9 @@
 ! Terms of use are as specified in LICENSE.txt
 
 #include "julienne-assert-macros.h"
-#include "assert_macros.h"
 
 submodule(layer_m) layer_s
-  use assert_m
-  use julienne_m, only : call_julienne_assert_, operator(.equalsExpected.)
+  use julienne_m, only : call_julienne_assert_, operator(.equalsExpected.), operator(.expect.)
   implicit none
 
 contains
@@ -21,7 +19,7 @@ contains
     line = adjustl(layer_lines(start)%string())
     hidden_layers = line == '['
     output_layer = line == '"output_layer": ['
-    call_assert(hidden_layers .or. output_layer)
+    call_julienne_assert(any([hidden_layers, output_layer]))
 
     layer%neuron = neuron_t(layer_lines, start+1)
     num_inputs = size(layer%neuron%weights())
@@ -54,7 +52,7 @@ contains
     line = adjustl(layer_lines(start)%string())
     hidden_layers = line == '['
     output_layer = line == '"output_layer": ['
-    call_assert(hidden_layers .or. output_layer)
+    call_julienne_assert(any([hidden_layers, output_layer]))
 
     layer%neuron = neuron_t(layer_lines, start+1)
     num_inputs = size(layer%neuron%weights())
