### 2025-10-21 cd80a58
refac(histogram_t): rm un{used,needed} getters
