### 2025-10-21 1100e4f
refac(histogram_t): store bin count, cardinality
This commit prepares for calculating histograms from multiple
data files by making histogram_t objects summable.  For this
purpose, histogram_t componnets now include bin count and
cardinality.  This commit replaces the bin_frequency_ component
with a bin_frequency() function that calculates frequency by
dividing the count by the cardinality.
