### 2025-10-21 bc23705
refac(histogram): constructor with preset range
This commit further prepares for computing aggregate histograms
from multi-file data by adding a histrogram_t user-defined
structure constructor that accepts a predefined range.  Doing so
will allow for using global extrema to define one range across
all files.
