### 2025-10-21 7bc90cc
feat(tensor-stats): read >1 training-data files
