### 2025-10-22 e3582ea
feat(tensor-stats): write 1 plot per training file
