+++ demo/app/tensor-statistics.F90
@@ -32,26 +32,12 @@ program tensor_statistics
   integer(int64) t_start, t_finish, clock_rate
   integer num_bins, start_step, stride
   integer, allocatable :: end_step
+  character(len=*), parameter :: data_json = "training_data_files.json", configuration_json = "training_configuration.json"
   logical raw
-  type(string_t), allocatable :: input_file_names(:), output_file_names(:)
 
   call system_clock(t_start, clock_rate)
   call get_command_line_arguments(num_bins, start_step, end_step, stride, raw)
-
-  associate( &
-     training_configuration => training_configuration_t(file_t("training_configuration.json")) &
-    ,training_data_files => training_data_files_t(file_t("training_data_files.json")) &
-  )
-
-    call compute_histograms( &
-       input_tensor_file_names  = training_data_files%fully_qualified_inputs_files() &
-      ,output_tensor_file_names = training_data_files%fully_qualified_outputs_files() &
-      ,time_data_file_name      = training_data_files%fully_qualified_time_file() &
-      ,input_component_names    = training_configuration%input_variable_names() &
-      ,output_component_names   = training_configuration%output_variable_names() &
-      ,raw = raw &
-    )
-  end associate
+  call compute_histograms(training_data_files_t(file_t(data_json)), training_configuration_t(file_t(configuration_json)), raw)
   call system_clock(t_finish)
 
   print '(a,g0)',"System clock time: ", real(t_finish - t_start, real64)/real(clock_rate, real64)
@@ -109,94 +95,123 @@ contains
  
   end subroutine get_command_line_arguments
 
-  subroutine compute_histograms( &
-    input_tensor_file_names, output_tensor_file_names, input_component_names, output_component_names, time_data_file_name, raw &
-  )
-    type(string_t), intent(in) :: input_tensor_file_names(:), output_tensor_file_names(:)
-    type(string_t), intent(in) :: input_component_names(:),   output_component_names(:), time_data_file_name
+  subroutine compute_histograms(training_data_files, training_configuration, raw)
+    type(training_data_files_t), intent(in) :: training_data_files
+    type(training_configuration_t), intent(in) :: training_configuration
     logical, intent(in) :: raw
 
     type(time_derivative_t), allocatable :: derivative(:)
-    type(NetCDF_variable_t) input_time, output_time
-    type(NetCDF_variable_t), allocatable :: input_variable(:), output_variable(:)
+    type(NetCDF_variable_t), allocatable :: input_variable(:,:), output_variable(:,:)
     type(NetCDF_file_t), allocatable :: NetCDF_input_file(:), NetCDF_output_file(:)
     double precision, allocatable, dimension(:) :: dt
     double precision, parameter :: tolerance = 1.E-07
     integer(int64) t_histo_start, t_histo_finish 
     integer t, t_end, v, f
+    type(string_t), allocatable :: input_tensor_file_names(:), output_tensor_file_names(:)
+    type(string_t), allocatable :: input_component_names(:), output_component_names(:)
+    character(len=:), allocatable :: time_data_file_name
 
-    allocate(input_variable(size(input_component_names)))
+    input_tensor_file_names  = training_data_files%fully_qualified_inputs_files()
+    output_tensor_file_names = training_data_files%fully_qualified_outputs_files()
+    time_data_file_name      = training_data_files%fully_qualified_time_file()
+    input_component_names    = training_configuration%input_variable_names()
+    output_component_names   = training_configuration%output_variable_names()
+
     allocate(NetCDF_input_file(size(input_tensor_file_names)))
+    allocate(input_variable(size(input_component_names), size(NetCDF_input_file)))
 
-    do f = 1, size(NetCDF_input_file)
[…]
