### 2025-10-22 5b352dd
feat(command-line): update flags & usage info
Usage:

./build/run-fpm.sh run tensor-statistics -- --bins <integer> \
  [--disaggregate] [--unified] [--start <integer>] [--end <integer>] [--stride <integer>]

where angular brackets (<>) denote user-provided values, square brackets denote optional
arguments, and flags have the following meanings:

--bins         number of histogram bins in which to collect variable values
--disaggregate generates one histogram plot file per variable per training-data file
--end          last time step
--start        first time step
--stride       interval betwee
[…]
