+++ demo/app/tensor-statistics.F90
@@ -22,43 +22,52 @@ program tensor_statistics
   use fiats_m, only : training_configuration_t, training_data_files_t
   implicit none
 
-  character(len=*), parameter :: usage =                new_line('') // new_line('') // & 
-    'Usage: '                                        // new_line('') // new_line('') // &
-    './build/run-fpm.sh run tensor-statistics -- --bins <integer> \' // new_line('') // &
-    '  [--raw] [--start <integer>] [--end <integer>] [--stride <integer>]'           // &
-                                                        new_line('') // new_line('') // &
-    'where angular brackets denote user-provided values and square brackets denote optional arguments.' // new_line('')
-
+  character(len=*), parameter :: usage = new_line('') &
+    // new_line('') // 'Usage: ' &
+    // new_line('') &
+    // new_line('') // './build/run-fpm.sh run tensor-statistics -- --bins <integer> \'  &
+    // new_line('') // '  [--disaggregate] [--unified] [--start <integer>] [--end <integer>] [--stride <integer>]' &
+    // new_line('') &
+    // new_line('') // 'where angular brackets (<>) denote user-provided values, square brackets denote optional' &
+    // new_line('') // 'arguments, and flags have the following meanings:' &
+    // new_line('') &
+    // new_line('') // '--bins         number of histogram bins in which to collect variable values' &
+    // new_line('') // '--disaggregate generates one histogram plot file per variable per training-data file' &
+    // new_line('') // '--end          last time step' &
+    // new_line('') // '--start        first time step' &
+    // new_line('') // '--stride       interval between time steps' &
+    // new_line('') // '--unified      maps all domains to [0,1] on one histogram plot per training-data file' &
+    // new_line('')
   integer(int64) t_start, t_finish, clock_rate
   integer num_bins, start_step, stride
   integer, allocatable :: end_step
   character(len=*), parameter :: data_json = "training_data_files.json", configuration_json = "training_configuration.json"
-  logical raw
+  logical disaggregated, unified
 
   call system_clock(t_start, clock_rate)
-  call get_command_line_arguments(num_bins, start_step, end_step, stride, raw)
-  call compute_histograms(training_data_files_t(file_t(data_json)), training_configuration_t(file_t(configuration_json)), raw)
+  call get_command_line_arguments(num_bins, start_step, end_step, stride, disaggregated, unified)
+  call compute_histograms(training_data_files_t(file_t(data_json)), training_configuration_t(file_t(configuration_json)), disaggregated)
   call system_clock(t_finish)
 
   print '(a,g0)',"System clock time: ", real(t_finish - t_start, real64)/real(clock_rate, real64)
   print *
   print '(a)',"______________________________________________________"
   print '(a)',"The *.plt files contain tensor ranges and histograms."
-  print '(a)',"If you have gnuplot installed, please execute the"
-  print '(a)',"following command to produce histogram visualizations:"
+  print '(a)',"If you have gnuplot installed, please execute a command"
+  print '(a)',"command to produce histogram visualizations:"
   print *
-  associate(file_name => trim(merge("plot-raw-histograms       ", "plot-normalized-histograms", raw)) // ".gnu")
-   print*,"  gnuplot app/" // file_name
+  associate(file_name => trim(merge("unified-", "        ", unified)) // "histogram-plot.gnu")
+    print*,"  gnuplot app/" // file_name
   end associate
   print *
   print '(a)',"_______________ tensor_statistics done________________"
 
 contains
 
-  subroutine get_command_line_arguments(num_bins, start_step, end_step, stride, raw)
[…]
