+++ demo/src/histogram_m.f90
@@ -23,21 +23,21 @@ module histogram_m
 
   interface histogram_t
 
-    pure module function construct(v, variable_name, num_bins, raw) result(histogram)
+    pure module function construct(v, variable_name, num_bins, disaggregated) result(histogram)
       implicit none
       real, intent(in) :: v(:,:,:,:)
       character(len=*), intent(in) :: variable_name
       integer, intent(in) :: num_bins
-      logical, intent(in) :: raw
+      logical, intent(in) :: disaggregated
       type(histogram_t) histogram
     end function
 
-    pure module function construct_in_range(variable_name, v, v_min, v_max, num_bins, raw) result(histogram)
+    pure module function construct_in_range(variable_name, v, v_min, v_max, num_bins, disaggregated) result(histogram)
       implicit none
       real, intent(in) :: v(:,:,:,:), v_min, v_max
       character(len=*), intent(in) :: variable_name
       integer, intent(in) :: num_bins
-      logical, intent(in) :: raw
+      logical, intent(in) :: disaggregated
       type(histogram_t) histogram
     end function
 
