### 2025-10-22 98f7124
chore(tensor-stats): rm unused args, update usage
Usage:

./build/run-fpm.sh run tensor-statistics -- --bins <integer> [[--disaggregate] | [--unified]]

where angular brackets (<>) denote user-provided values, square brackets ([]) denote optional
arguments, a pipe (|) separates exclusive alternatives, and flags have the following meanings:

--bins         number of histogram bins in which to collect variable values
--disaggregate generates one histogram plot file per variable per training-data file
--unified      maps all domains to [0,1] on one histogram plot per training-data file
