### 2025-10-22 dc4e63f
refac(tensor-stats): encapsulate command-line args
