+++ demo/app/tensor-statistics.F90
@@ -11,7 +11,6 @@ program tensor_statistics
 
   ! External dependencies:
   use julienne_m, only : command_line_t, file_t, string_t, call_julienne_assert_
-  use ieee_arithmetic, only : ieee_is_nan
   use iso_fortran_env, only : int64, real64
     
   ! Internal dependencies:
@@ -34,208 +33,205 @@ program tensor_statistics
     // new_line('') // '--disaggregate generates one histogram plot file per variable per training-data file' &
     // new_line('') // '--unified      maps all domains to [0,1] on one histogram plot per training-data file' &
     // new_line('')
+
+  type command_line_arguments_t
+    integer :: num_bins = 0
+    logical disaggregated, unified
+  end type
   integer(int64) t_start, t_finish, clock_rate
-  integer num_bins
-  character(len=*), parameter :: data_json = "training_data_files.json", configuration_json = "training_configuration.json"
-  logical disaggregated, unified
 
-  call system_clock(t_start, clock_rate)
-  call get_command_line_arguments(num_bins, disaggregated, unified)
-  call compute_histograms(training_data_files_t(file_t(data_json)), training_configuration_t(file_t(configuration_json)), disaggregated)
-  call system_clock(t_finish)
-
-  print '(a,g0)',"System clock time: ", real(t_finish - t_start, real64)/real(clock_rate, real64)
-  print *
-  print '(a)',"______________________________________________________"
-  print '(a)',"The *.plt files contain tensor ranges and histograms."
-  print '(a)',"If you have gnuplot installed, please execute a command"
-  print '(a)',"command to produce histogram visualizations:"
-  print *
-  associate(file_name => trim(merge("unified-", "        ", unified)) // "histogram-plot.gnu")
-    print*,"  gnuplot app/" // file_name
+  associate(args => command_line_arguments())
+    if (args%num_bins < 1) error stop usage
+    call system_clock(t_start, clock_rate)
+    call calc_write_stats(args, training_data_files_t(file_t("training_data_files.json")), training_configuration_t(file_t("training_configuration.json")))
+    call system_clock(t_finish)
+    call print_trailer(args, elapsed_time = real(t_finish - t_start, real64)/real(clock_rate, real64))
   end associate
-  print *
-  print '(a)',"_______________ tensor_statistics done________________"
 
 contains
 
-  subroutine get_command_line_arguments(num_bins, disaggregated, unified)
-    integer, intent(out) :: num_bins
-    logical, intent(out) :: disaggregated, unified
-
-    ! local variables
+  type(command_line_arguments_t) function command_line_arguments()
     type(command_line_t) command_line
     character(len=:), allocatable :: bins_string
-
+    command_line_arguments%disaggregated = command_line%argument_present(["--disaggregated"])
+    command_line_arguments%unified = command_line%argument_present(["--unified"])
     bins_string = command_line%flag_value("--bins")
-    disaggregated = command_line%argument_present(["--disaggregated"])
-    unified = command_line%argument_present(["--unified"])
+    if (len(bins_string) /= 0) read(bins_string,*) command_line_arguments%num_bins
+  end function
 
-    associate(required_arguments => len(bins_string)/=0)
-       if (.not. required_arguments) error stop usage 
+  subroutine print_trailer(args, elapsed_time)
+    type(command_line_arguments_t), intent(in) :: args
+    real(real64), intent(in) :: elapsed_time
+
+    print '(a,g0)',"System clock time: ", elapsed_time
+    print *
+    print '(a)',"______________________________________________________"
+    print '(a)',"The *.plt files contain tensor ranges and histograms."
+    print '(a)',"If you have gnuplot installed, please execute a command"
+    print '(a)',"command to produce histogram visualizations:"
+    print *
[…]
