### 2025-10-25 c6bdba0
test(histogram_t): check uniform distribution
