### 2025-10-27 aa05968
chore(demo/histogram): rm optional arguments
