+++ demo/src/NetCDF_variable_s.F90
@@ -68,7 +68,7 @@ contains
   module procedure default_real_histogram
     select case(self%rank())
     case (4)
-      histogram = histogram_t(self%values_4D_ , self%name_, num_bins, raw)
+      histogram = histogram_t(self%values_4D_ , self%name_, num_bins)
     case default
       error stop 'NetCDF_variable_s(default_real_histogram): unsupported rank'
     end select 
