### 2025-10-27 2a264b0
refac(histogram_t): match constructor arg order
