### 2025-10-27 78bbd5e
feat(demo/histogram): rm domain normalization
