### 2025-10-27 4966a08
chore(histogram_s): blank space eds (indentation)
