### 2025-10-27 ed7dd04
feat(histogram): default(none) locality
