### 2025-10-27 bec7f65
refac(histogram): map constructor inputs to 1D
