### 2025-10-27 145a75d
feat(NetCDF_variable): 1-3D histogram construction
