### 2025-10-27 f7ef70b
test(histogram): construct from bracketed array(s)
This commit tests behavior of the histogram_t constructor when the
first actual argument is one or more bracketed arrays such as [v]
or [u,v,w], where u, v, and w are each a multidimensional array.
