### 2025-10-27 703c9c9
feat(NetCDF_var): aggregate histograms across files
