### 2025-10-27 40799c0
WIP(demo/tensor-statistics): calc derivs
