### 2025-10-27 ec54158
fix(NetCDF_variable_s,histogram_s): assertions
