### 2025-10-27 5e16f57
fix(tensor-statistics): alloc derivs just once
