### 2025-10-30 6fc16ba
fix(tensor-statistics): refomat timer output
