### 2025-10-30 c511fae
doc(joss): rm all figures except Fig. 1
