### 2025-10-30 168431d
Cut down discussion of Fortran standard in Statement of Need
