### 2025-10-30 d388378
doc(joss): reduce/tailor Statement of Need
