### 2025-10-31 010ef55
WIP: augmentation process hangs
