### 2025-10-31 #8785 @espottesmith
Good point on "native". I would support the authors changing that use of term.
