+++ doc/joss/paper.bib
@@ -3,8 +3,8 @@
   author={Rouson, Damian and Bai, Zhe and Bonachea, Dan and Dibba, Baboucarr and Gutmann, Ethan and Rasmussen, Katherine and Torres, David and Welsman, Jordan and Zhang, Yunhao},
   booktitle={2025 Improving Scientific Software Conference},
   organization={University Corporation for Atmospheric Research},
-  pages={(in review)},
-  note = "\url{https://ucar-sea.github.io/SEA-ISS-2025-Cloud-microphysics-training}",
+  doi = "10.25344/S4QS3J",
+  _note = "\url{https://ucar-sea.github.io/SEA-ISS-2025-Cloud-microphysics-training}",
   year={2025}
 }
 
@@ -43,7 +43,7 @@
 @inproceedings{curcic2019parallel,
   title={A parallel {F}ortran framework for neural networks and deep learning},
   author={Curcic, Milan},
-  booktitle={Acm sigplan fortran forum},
+  booktitle={ACM SIGPLAN Fortran Forum},
   volume={38},
   number={1},
   pages={4--21},
