### 2025-10-31 27b8344
"native" => "built-in"
