### 2025-11-02 8d1d2d6
More citation updates
