+++ doc/joss/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{rouson2025cloud,
   title={Cloud microphysics training and aerosol inference with the {F}iats deep learning library},
   author={Rouson, Damian and Bai, Zhe and Bonachea, Dan and Dibba, Baboucarr and Gutmann, Ethan and Rasmussen, Katherine and Torres, David and Welsman, Jordan and Zhang, Yunhao},
-  booktitle={2025 Improving Scientific Software Conference},
+  booktitle={Improving Scientific Software Conference (ISS25)},
   organization={University Corporation for Atmospheric Research},
   doi = "10.25344/S4QS3J",
   _note = "\url{https://ucar-sea.github.io/SEA-ISS-2025-Cloud-microphysics-training}",
@@ -29,10 +29,10 @@
   pages={7895}
 }
 
-@techreport{rasmussen2022agile,
+@inproceedings{rasmussen2022agile,
   title={Agile Acceleration of LLVM Flang Support for {F}ortran 2018 Parallel Programming},
   author={Rasmussen, Katherine and Rouson, Damian and George, Naj{\'e} and Bonachea, Dan and Kadhem, Hussain and Friesen, Brian},
-  institution  = {Lawrence Berkeley National Laboratory (LBNL), Berkeley, CA (United States)},
+  booktitle={Research Poster at the International Conference for High Performance Computing, Networking, Storage, and Analysis (SC22)},
   doi={10.25344/S4CP4S},
   url={https://www.osti.gov/biblio/1886637},
   year={2022},
