### 2025-11-02 919ed27
Rephrase trainable_network_t dependence on file_t
