### 2025-11-04 cbc5b0d
Minor cleanups in bibliography
