+++ doc/joss/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{rouson2025cloud,
   title={Cloud microphysics training and aerosol inference with the {F}iats deep learning library},
   author={Rouson, Damian and Bai, Zhe and Bonachea, Dan and Dibba, Baboucarr and Gutmann, Ethan and Rasmussen, Katherine and Torres, David and Welsman, Jordan and Zhang, Yunhao},
-  booktitle={Improving Scientific Software Conference (ISS25)},
+  booktitle={Improving Scientific Software Conference (ISS)},
   organization={University Corporation for Atmospheric Research},
   doi = "10.25344/S4QS3J",
   _note = "\url{https://ucar-sea.github.io/SEA-ISS-2025-Cloud-microphysics-training}",
@@ -32,7 +32,7 @@
 @inproceedings{rasmussen2022agile,
   title={Agile Acceleration of LLVM Flang Support for {F}ortran 2018 Parallel Programming},
   author={Rasmussen, Katherine and Rouson, Damian and George, Naj{\'e} and Bonachea, Dan and Kadhem, Hussain and Friesen, Brian},
-  booktitle={Research Poster at the International Conference for High Performance Computing, Networking, Storage, and Analysis (SC22)},
+  booktitle={Research Poster at the International Conference for High Performance Computing, Networking, Storage, and Analysis (SC)},
   doi={10.25344/S4CP4S},
   url={https://www.osti.gov/biblio/1886637},
   year={2022},
@@ -76,7 +76,7 @@
 @inproceedings{rouson2022caffeine,
   title = {{Caffeine: CoArray Fortran Framework of Efficient Interfaces to Network Environments}},
   author={Rouson, Damian and Bonachea, Dan},
-  booktitle={2022 IEEE/ACM Eighth Workshop on the LLVM Compiler Infrastructure in HPC (LLVM-HPC)},
+  booktitle={Eighth Workshop on the LLVM Compiler Infrastructure in HPC (LLVM-HPC)},
   pages={34--42},
   year={2022},
   doi={10.25344/S4459B},
@@ -86,7 +86,7 @@
 @inproceedings{bonachea2024prif,
   title={{Parallel Runtime Interface for Fortran (PRIF): A Multi-Image Solution for LLVM Flang}},
   author={Bonachea, Dan and Rasmussen, Katherine and Richardson, Brad and Rouson, Damian},
-  booktitle={Tenth Workshop on the LLVM Compiler Infrastructure in HPC (LLVM-HPC2024)},
+  booktitle={Tenth Workshop on the LLVM Compiler Infrastructure in HPC (LLVM-HPC)},
   organization={IEEE},
   year={2024},
   doi={10.25344/S4N017},
@@ -96,7 +96,7 @@
 @inproceedings{bonachea2025llvm,
   title={Lowering and Runtime Support for {Fortran's} Multi-Image Parallel Features using {LLVM Flang}, {PRIF}, and {Caffeine}},
   author={Bonachea, Dan and Rasmussen, Katherine and Rouson, Damian and Pailleux, Jean-Didier and Renault, Etienne and Richardson, Brad},
-  booktitle={Eleventh Workshop on the LLVM Compiler Infrastructure in HPC (LLVM-HPC2025)},
+  booktitle={Eleventh Workshop on the LLVM Compiler Infrastructure in HPC (LLVM-HPC)},
   organization={ACM},
   year={2025},
   doi={10.25344/S4G883},
@@ -105,7 +105,7 @@
 @techreport{prif-0.5,
   author       = {Bonachea, Dan and Rasmussen, Katherine and Richardson, Brad and Rouson, Damian},
   title        = {Parallel Runtime Interface for {F}ortran (PRIF) Specification (Rev. 0.5)},
[…]
