### 2025-11-06 1884cd2
doc(README.md): update compiler flags
