### 2025-11-06 4e76dbc
rm training code
