### 2025-11-09 #8785 @bonachea
@espottesmith author tasks complete:

    * **[version 1.2.0](https://github.com/BerkeleyLab/fiats/releases/tag/1.2.0)**
    * **<https://doi.org/10.5281/zenodo.17566910>**
