### 2025-11-10 #8785 @espottesmith
Small quirk: on Zenodo, I'm seeing the version as "1,2,0", rather than 1.2.0. Can you confirm, @bonachea?
