### 2025-11-10 c4d35ec
Fix a mangled DOI
This DOI was still resolving, but it's not canonical and editorialbot
was presumably rejecting it for this reason.
