+++ doc/joss/paper.bib
@@ -70,7 +70,7 @@
   author = {Robert W. Numrich},
   publisher ={{CRC Press}},
   year = {2018},
-  doi={doi:10.1201/9780429437182}
+  doi={10.1201/9780429437182}
 }
 
 @inproceedings{rouson2022caffeine,
