### 2025-11-10 #8785 @bonachea
@espottesmith
> Small quirk: on Zenodo, I'm seeing the version as "1,2,0", rather than 1.2.0.

Thanks for the report, typo fixed.

Re: "Reference check summary":
I fixed the "invalid" DOI, which was really just editorialbot choking on a non-canonical (but functional) DOI.

On the "missing" DOI, as previously explained on Aug 12:

> ? MISSING DOIs
> - 10.1109/llvm-hpc56686.2022.00009 may be a valid DOI for title: Caffeine: CoArray Fortran Framework of Efficient I...

This is a false report: the DOI is **not** "missing".
The reference in question (Rouson, D., & Bonachea, D. (2022)) already includes an open-access DOI: <https://doi.org/10.25344/S4459B>
editorialbot is suggesting the pay-walled publisher DOI, but the open-access DOI provides superior availability.
