### 2025-11-11 #8785 @espottesmith
Cool. Thanks for confirming. I'm very happy with you keeping the open-access DOI.
