### 2025-11-12 28b2324
doc(joss): make edits based on review
