### 2025-11-12 #8785 @rouson
@espottesmith we have made changes that we believe address each of your concerns except those regarding Figure 1.  I will investigate whether the Mermaid tool that created that figure allows for customization.  Other options are we could leave the figure as is or we could rotate the figure 90 degrees or we could remove the figure (in which case, it probably makes sense to remove most or all of the subsequent text).  Please let us know your preference.
