### 2025-11-13 #8785 @espottesmith
I think messing with Mermaid to customize the figure would be ideal. Rotating 90 degrees would be fine if that isn't possible.
