### 2025-11-22 48cc328
Rotate figure 1
This matches the figure shape closer to the page aspect ratio,
allowing the fonts to be larger and more readable when printed

Also need to relocate it within the markdown, since this tool chain
does not seem to support floating figures.
