### 2025-11-22 #8785 @bonachea
Hi @espottesmith - We made some attempts to get Mermaid to rearrange the figure, but the tool seems to lack the control capabilities needed to make that work. So we've instead opted to rotate the figure to increase the font size as you requested.
