### 2025-12-04 cddf966
doc(joss): delete authors' related work, ed need
This commit
1. Removes related work from the Summary and
2. Adds a sentence to the Statement of Need.
