### 2025-12-05 #8785 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
