### 2025-12-06 #8785 @espottesmith
The LBNL-BSD license has been accepted before; see my previous JOSS submission [RNMC](https://github.com/BlauGroup/RNMC/blob/main/LICENSE).
