### 2025-12-06 #8785 @crvernon
Thanks for the clarification @espottesmith - given this, I'll proceed with acceptance.
